refactor: use runtime context for settings
