+++ src/atlas_assistant/context.py
@@ -1,0 +1,10 @@
+from dataclasses import dataclass
+
+from .settings import Settings
+
+
+@dataclass
+class Context:
+    """Immutable values shared between tools"""
+
+    settings: Settings
