+++ src/atlas_assistant/dataset.py
@@ -2,12 +2,7 @@
 
 from __future__ import annotations
 
-from langchain.messages import ToolMessage
-from langchain.tools import ToolRuntime, tool
-from langgraph.types import Command
 from pydantic import BaseModel
-
-from .settings import Settings, get_settings
 
 
 class Item(BaseModel):
@@ -110,44 +105,3 @@ class Metadata(BaseModel):
         return Dataset(
             item=Item.model_validate_json(self.item), asset_key=self.asset_key
         )
-
-
-class SearchResult(BaseModel):
-    """The result of an embeddings search"""
-
-    dataset: Dataset
-    """The dataset"""
-
-    score: float
-    """The search score"""
-
-
-@tool
-def select_dataset(query: str, runtime: ToolRuntime) -> Command[None]:
-    """Selects a dataset based on a user's query.
-
-    Args:
-        query: Search terms to select the dataset
-    """
-    settings = get_settings()
-    search_result = search(query, settings)
-    content = f"Selected dataset: {search_result.dataset.get_description()}"
-    return Command(
-        update={
-            "messages": [
-                ToolMessage(
-                    content=content,
-                    tool_call_id=runtime.tool_call_id,
-                )
-            ],
-            "dataset": search_result.dataset,
-        }
-    )
-
-
-def search(query: str, settings: Settings) -> SearchResult:
-    """Search the embeddings for datasets that match the query"""
-    embeddings = settings.get_embeddings()
-    results = embeddings.similarity_search_with_score(query, k=1)
-    dataset = Metadata.model_validate(results[0][0].metadata).to_dataset()
-    return SearchResult(dataset=dataset, score=results[0][1])
