+++ tests/test_agent.py
@@ -3,6 +3,7 @@ from langchain_core.messages import HumanMessage
 
 import atlas_assistant.agent
 from atlas_assistant.agent import Agent
+from atlas_assistant.context import Context
 from atlas_assistant.settings import Settings
 
 pytestmark = pytest.mark.integration
@@ -13,10 +14,11 @@ def agent(settings: Settings) -> Agent:
     return atlas_assistant.agent.create_agent(settings)
 
 
-def test_kenya_crops(agent: Agent) -> None:
+def test_kenya_crops(agent: Agent, settings: Settings) -> None:
     _ = agent.invoke(
         {
             "messages": [HumanMessage("What crops are being grown in Kenya?")],
         },
         config={"configurable": {"thread_id": "test"}},
+        context=Context(settings=settings),
     )
